Add soft-delete support and system logs
Introduce soft-delete (deletedAt) across many Prisma models and add a SystemLog model. Add a logger helper (src/lib/logger.ts) and wire system logs route. Update auth middleware to treat soft-deleted users as unauthorized. Replace hard deletes with updates that set deletedAt and add deletedAt filters to queries in admin and evaluation routes. Add createLog calls for admin actions (create/update/delete) and refactor events routing (old src/routes/events.ts removed; new modular routes under src/routes/events/). Misc: import adjustments in src/index.ts and various Prisma schema formatting/field updates.
