+++ src/routes/systemlogs.ts
@@ -1,0 +1,79 @@
+import { Hono } from "hono";
+import { zValidator } from "@hono/zod-validator";
+import { z } from "zod";
+import { authMiddleware } from "../middlewares/auth.js";
+import { adminOnly } from "../middlewares/adminOnly.js";
+import { prisma } from "../lib/prisma.js";
+
+const systemLogs = new Hono();
+
+const logsQuerySchema = z.object({
+  page: z.string().optional().transform((val) => (val ? parseInt(val, 10) : 1)),
+  limit: z.string().optional().transform((val) => (val ? parseInt(val, 10) : 20)),
+  search: z.string().optional(),
+  action: z.string().optional(),
+  userId: z.string().optional(),
+  startDate: z.string().optional(),
+  endDate: z.string().optional(),
+});
+
+systemLogs
+  .use("*", authMiddleware, adminOnly)
+  .get("/", zValidator("query", logsQuerySchema), async (c) => {
+    const { page, limit, search, action, userId, startDate, endDate } = c.req.valid("query");
+    const skip = (page - 1) * limit;
+
+    const where: any = {};
+
+    if (search) {
+      where.OR = [
+        { action: { contains: search, mode: "insensitive" } },
+        { details: { contains: search, mode: "insensitive" } },
+        { user: { username: { contains: search, mode: "insensitive" } } },
+        { user: { email: { contains: search, mode: "insensitive" } } },
+      ];
+    }
+
+    if (action) where.action = action;
+    if (userId) where.userId = userId;
+    
+    if (startDate || endDate) {
+      where.createdAt = {};
+      if (startDate) where.createdAt.gte = new Date(startDate);
+      if (endDate) where.createdAt.lte = new Date(endDate);
+    }
+
+    where.deletedAt = null;
+
+    const [total, logs] = await Promise.all([
+      prisma.systemLog.count({ where }),
+      prisma.systemLog.findMany({
+        where,
+        take: limit,
+        skip,
+        orderBy: { createdAt: "desc" },
+        include: {
+          user: {
+            select: {
+              id: true,
+              username: true,
+              email: true,
+              image: true,
+            },
+          },
+        },
+      }),
+    ]);
+
+    return c.json({
+      data: logs,
+      meta: {
+        total,
+        page,
+        limit,
+        totalPages: Math.ceil(total / limit),
+      },
+    });
+  });
+
+export default systemLogs;
